fix: remove built-in $authentication store
This feature was intended to be an easier API for a built-in
hybrid message[^1], but turned out to be problematic. By design
the store will emit its value on the message bus if the message
bus has no previous message for the channel and topic.

For authentication in a hybrid webview there is no client-side
initial value (it's an Authorization HTTP request header). Unless
a client-side script primes the message bus with a fake message
on "system/authentication", this store ends up sending a signal
to the native side that the user has logged out (authentication
token is null).

Since the method to detect login state and prime the message
bus will be different from case to case, this built-in store
is removed.

BREAKING CHANGE: $authentication is removed

[1]: https://podium-lib.io/docs/guides/hybrid#message-contracts